% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grec-main.R
\docType{data}
\name{bathy}
\alias{bathy}
\title{Bathymetric data}
\format{A \code{list} with bathymetric information from ETOPO source.}
\usage{
bathy
}
\description{
Bathymetric maps downloaded from ERDDAP for running examples with \code{grec} functions.
}
\references{
ERDDAP website: \url{https://coastwatch.pfeg.noaa.gov/erddap/index.html}
}
