% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grec-main.R
\docType{data}
\name{colPalette}
\alias{colPalette}
\title{Default color palette most using on environmental representations.}
\format{A vector of 2000 colors in RGB format.}
\usage{
colPalette
}
\description{
Vector with 2000 colors generated from \code{tim.colors} function.
}
\references{
\code{fields} package
}
