% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grec-main.R
\docType{data}
\name{exampleSSTData}
\alias{exampleSSTData}
\title{Example map of SST from Aqua MODIS}
\format{A \code{list} with 3 levels (x, y, z). \code{x} and \code{y} are numeric vectors, \code{z}
is a numeric \code{matrix}.}
\usage{
exampleSSTData
}
\description{
A \code{list} with 3 levels (x, y, z) containing information of an SST map: longitud, latitude and
values (as a \code{matrix}). Info was downloaded from Aqua MODIS, March, 2010.
}
