% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grec-main.R
\name{extraParams}
\alias{extraParams}
\title{Gets the extra parameters for \code{grec} functions}
\usage{
extraParams(fx = c("detectFronts"))
}
\arguments{
\item{fx}{\code{character} vector indicating the values that will be returned.}
}
\value{
A \code{list} with the extra parameters used by default.
}
\description{
Show as a list the extra parameters for main \code{grec} function.
}
\details{
This function is usefull for advance users that require to modify intermediate steps in terms of
filter application.
}
\examples{
# For getting all the extra parameters
extraParams()
}
