% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grec-main.R
\docType{package}
\name{grec-package}
\alias{grec-package}
\alias{grec}
\title{GRadient-based RECognition of spatial patterns in environmental data}
\description{
Provides algorithms for detecting spatial patterns from 2-D
oceanographic data using image processing methods based on Gradient
Recognition.
}
\author{
Wencheng Lau-Medrano, \email{luis.laum@gmail.com}
}
\concept{environmental-data}
\concept{gradient}
\concept{pattern-detection}
