% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grec-main.R
\docType{data}
\name{grecExData}
\alias{grecExData}
\alias{exampleSSTData}
\title{Example map of SST from Aqua MODIS}
\format{A \code{list} with 3 levels (x, y, z). \code{x} and \code{y} are numeric vectors, \code{z}
is a numeric \code{matrix}.}
\usage{
exampleSSTData
}
\description{
A \code{list} containing three environmental dataset for using as examples. Datasets
of Surface Sea Temperature (\code{sst}), Chlorophyll at surface (\code{chlorophyll}) and Bathymetry
(topography) were dowloaded from ERDDAP website.

Chlorophyll and SST data are XYZT lists with monthly info from February to April 2015 from source
Aqua MODIS, March, 2010.
}
\references{
ERDDAP website: \url{https://coastwatch.pfeg.noaa.gov/erddap/index.html}
}
