% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grec-main.R
\docType{data}
\name{NW_USA_SST}
\alias{NW_USA_SST}
\title{Sea Surface Temperature Data for NW USA}
\format{A \code{RasterLayer} object with SST information from Aqua MODIS source.}
\usage{
NW_USA_SST
}
\description{
SST RasterLayer map for running examples with \code{grec} functions.
}
\references{
ERDDAP website: \url{https://coastwatch.pfeg.noaa.gov/erddap/index.html}
}
