% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grec-main.R
\docType{data}
\name{sst}
\alias{sst}
\title{Sea Surface Temperature Data}
\format{
A \code{list} with SST information from February to April of Aqua
MODIS source.
}
\usage{
sst
}
\description{
SST maps downloaded from ERDDAP for running examples with
\code{grec} functions.
}
\references{
ERDDAP website: \url{https://coastwatch.pfeg.noaa.gov/erddap/index.html}
}
\keyword{datasets}
