% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grec-main.R
\name{detectFronts}
\alias{detectFronts}
\title{Apply gradient-based methodologies to environmental data}
\usage{
detectFronts(...)
}
\arguments{
\item{...}{Same arguments than \link[grec]{getGradients}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function empowers users to analyze data from various sources,
including numeric \code{matrix}, \code{array}s, XYZ-\code{list}s,
\code{SpatRaster}s, or \code{RasterLayer}s*, by applying gradient-seeking
methodologies.
}
\details{
Since version 1.6.0, this function has been entirely replaced by
\link[grec]{getGradients}. As of version 2.0.0, \code{detectFronts} will no longer
be available.
}
