% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grec-main.R
\docType{package}
\name{grec-package}
\alias{grec-package}
\alias{grec}
\title{Classification of spatial patterns from environmental data through GRadient RECognition}
\description{
It provides algorithms for detection of spatial patterns from oceanographic data.
}
\author{
Wencheng Lau-Medrano, \email{luis.laum@gmail.com}
}
\keyword{environmental-data}
\keyword{gradient,}
\keyword{pattern-detection,}
