% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grec-package.R
\docType{package}
\name{grec-package}
\alias{grec}
\alias{grec-package}
\title{GRadient-based RECognition of spatial patterns in environmental data}
\description{
Provides algorithms for detecting spatial patterns from 2-D
oceanographic data using image processing methods based on Gradient
Recognition.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LuisLauM/grec}
  \item Report bugs at \url{https://github.com/LuisLauM/grec/issues}
}

}
\author{
\strong{Maintainer}: Wencheng Lau-Medrano \email{luis.laum@gmail.com}

}
\keyword{internal}
