% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grec-main.R
\docType{data}
\name{NW_USA_SST}
\alias{NW_USA_SST}
\title{Sea Surface Temperature Data}
\format{A \code{list} with SST information for a single day: 2012-10-13, GHRSST, 1km.}
\usage{
NW_USA_SST
}
\description{
SST map for Northwest USA, 77W-63W & 34N-45N.
}
\references{
Download source: \url{https://podaac.jpl.nasa.gov/GHRSST}
}
